Packaging_Dog_Breed_ML/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.2.0=eigen
abseil-cpp=20211102.0=he9d5cce_0
absl-py=2.1.0=py39hecd8cb5_0
aiohttp=3.9.5=py39h6c40b1e_0
aiosignal=1.2.0=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.3=py39hecd8cb5_0
attrs=23.1.0=py39hecd8cb5_0
blas=1.0=openblas
blinker=1.6.2=py39hecd8cb5_0
bottleneck=1.3.7=py39h7b7cdfe_0
brotli-python=1.0.9=py39hcec6c5f_8
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2024.3.11=hecd8cb5_0
cachetools=5.3.3=py39hecd8cb5_0
certifi=2024.2.2=py39hecd8cb5_0
cffi=1.16.0=py39h6c40b1e_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39hecd8cb5_0
cryptography=41.0.3=py39ha2381d6_0
flatbuffers=2.0.0=h23ab428_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.4.0=py39h6c40b1e_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h6c40b1e_3
google-auth=2.29.0=py39hecd8cb5_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.46.1=h64d96ca_1
grpcio=1.42.0=py39ha29bfda_0
h5py=3.11.0=py39h8cd6c6c_0
hdf5=1.12.1=h2b2ad87_2
icu=68.1=h23ab428_0
idna=3.7=py39hecd8cb5_0
importlib-metadata=7.0.1=py39hecd8cb5_0
joblib=1.4.0=py39hecd8cb5_0
jpeg=9e=h6c40b1e_1
keras=2.11.0=py39_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
krb5=1.20.1=hdba6334_1
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libcurl=8.2.1=ha585b31_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libedit=3.1.20230828=h6c40b1e_0
libev=4.33=h9ed2024_1
libffi=3.4.4=hecd8cb5_1
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libnghttp2=1.52.0=h1c88b7d_1
libopenblas=0.3.21=h54e7dc3_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=3.20.3=hfff2838_0
libssh2=1.10.0=hdb2fb19_2
libtiff=4.5.1=hcec6c5f_0
libwebp=1.3.2=hf6ce154_0
libwebp-base=1.3.2=h6c40b1e_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_1
markdown=3.4.1=py39hecd8cb5_0
markupsafe=2.1.3=py39h6c40b1e_0
multidict=6.0.4=py39h6c40b1e_0
ncurses=6.4=hcec6c5f_0
numexpr=2.8.7=py39hf6dca73_0
numpy=1.26.4=py39hf6dca73_0
numpy-base=1.26.4=py39hd8f4981_0
oauthlib=3.2.2=py39hecd8cb5_0
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1w=hca72f7f_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=23.2=py39hecd8cb5_0
pandas=2.2.1=py39hcec6c5f_0
pillow=10.0.1=py39h7d39338_0
pip=24.0=py39hecd8cb5_0
protobuf=3.20.3=py39hcec6c5f_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pybind11-abi=4=hd3eb1b0_1
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.8.0=py39hecd8cb5_0
pyopenssl=23.2.0=py39hecd8cb5_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.18=h218abb5_0
python-dateutil=2.9.0post0=py39hecd8cb5_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py39hecd8cb5_0
re2=2022.04.01=he9d5cce_0
readline=8.2=hca72f7f_0
requests=2.31.0=py39hecd8cb5_1
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.4.2=py39h3ea8b11_1
scipy=1.13.0=py39ha516a68_0
setuptools=69.5.1=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=hcec6c5f_1
sqlite=3.45.3=h6c40b1e_0
tensorboard=2.11.0=py39_0
tensorboard-data-server=0.6.1=py39h7242b5c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.11.0=eigen_py39hcfa9fec_0
tensorflow-base=2.11.0=eigen_py39hbf87084_0
tensorflow-estimator=2.11.0=py39_0
termcolor=2.1.0=py39hecd8cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.14=h4d00af3_0
typing_extensions=4.11.0=py39hecd8cb5_0
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py39hecd8cb5_1
werkzeug=2.3.8=py39hecd8cb5_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.14.1=py39hca72f7f_0
xz=5.4.6=h6c40b1e_1
yarl=1.9.3=py39h6c40b1e_0
zipp=3.17.0=py39hecd8cb5_0
zlib=1.2.13=h4b97444_1
zstd=1.5.5=hc035e20_2
